+++ src/pages/App.jsx
@@ -1,0 +1,11 @@
+import '../utils/style/App.css';
+
+function App() {
+
+  return (
+    <div className="App">
+    </div>
+  );
+}
+
+export default App;
